backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pdfplumber
Pillow
langchain
chromadb
openai
streamlit
sentence-transformers
